requirements.txt
PyYAML
cloudflare
maxcdn
